Moved Rings to a member variable/attribute of the readoutgenerator that can be defined in the constructor as needed instead if being passed to makePackets and on to generateData
