+++ src/common/readout/vmm3/generators/CMakeLists.txt
@@ -16,3 +16,19 @@ set(udpgen_vmm3_freia_SRC
   )
 create_executable(udpgen_vmm3_freia)
 target_compile_definitions(udpgen_vmm3_freia PUBLIC FREIA_GENERATOR)
+
+
+set(udpgen_vmm3_cspec_INC
+  ../VMM3Parser.h
+  ../generators/ReadoutGeneratorBase.h
+  ${ESS_MODULE_DIR}/cspec/generators/ReadoutGenerator.h
+  ${ESS_SOURCE_DIR}/common/testutils/DataFuzzer.h
+  )
+set(udpgen_vmm3_cspec_SRC
+  udpgen_vmm3_main.cpp
+  ../generators/ReadoutGeneratorBase.cpp
+  ${ESS_MODULE_DIR}/cspec/generators/ReadoutGenerator.cpp
+  ${ESS_SOURCE_DIR}/common/testutils/DataFuzzer.cpp
+  )
+create_executable(udpgen_vmm3_cspec)
+target_compile_definitions(udpgen_vmm3_cspec PUBLIC CSPEC_GENERATOR)
