+++ src/modules/freia/generators/ReadoutGenerator.h
@@ -18,8 +18,8 @@ namespace Freia {
 class ReadoutGenerator : public ReadoutGeneratorBase {
   public:
    using ReadoutGeneratorBase::ReadoutGeneratorBase;
-  private:
-  void generateData(uint8_t Rings, uint16_t NumReadouts) override;
+  protected:
+  void generateData(uint16_t NumReadouts) override;
    const uint32_t TimeToFirstReadout{1000};
 }; 
 }
